requirements.txt
aiofiles==24.1.0
alembic==1.15.1
doctran==0.0.14
Flask==3.1.0
flask_cors==5.0.1
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
injector==0.22.0
jieba==0.42.1
langchain==0.3.22
langchain_cohere==0.4.3
langchain_community==0.3.20
langchain_core==0.3.49
langchain_deepseek==0.1.3
langchain_experimental==0.3.4
langchain_huggingface==0.1.2
langchain_openai==0.3.11
langchain_pinecone==0.2.3
langchain_text_splitters==0.3.7
langchain_weaviate==0.0.4
langgraph==0.3.21
marshmallow==3.26.1
numpy==2.2.4
openai==1.70.0
pandas==2.2.3
pydantic==2.11.1
pytest==8.3.5
python-dotenv==1.1.0
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.6.1
tiktoken==0.9.0
umap_learn==0.5.7
weaviate_client==4.11.3
WTForms==3.2.1
